完成二课核心功能

+++ app/src/main/java/com/thryan/secondclass/core/result/ScoreInfo.kt
@@ -1,3 +1,14 @@
 package com.thryan.secondclass.core.result
 
-data class ScoreInfo(val score: Int, val item: Int, val integrity_value: Int,val activity: Int)
+import java.math.BigDecimal
+
+/**
+ * @param score 积分
+ * @param item 申报活动数？
+ * @param integrity_value 诚信值
+ * @param activity 完成活动数
+ */
+data class ScoreInfo(
+    val score: BigDecimal, val item: Int, val integrity_value: Int,
+    val activity: Int
+)
